src/third_party/blink/web_tests/http/tests/security/isolatedWorld/bypass-main-world-csp-for-xhr-expected.txt

CONSOLE ERROR: line 113: Refused to connect to 'http://localhost:8000/security/isolatedWorld/resources/cross-origin-xhr.txt' because it violates the following Content Security Policy directive: "connect-src 'none'".

CONSOLE ERROR: Refused to connect to 'http://localhost:8000/security/isolatedWorld/resources/cross-origin-xhr.txt' because it violates the following Content Security Policy directive: "connect-src 'none'".

CONSOLE ERROR: Refused to connect to 'http://localhost:8000/security/isolatedWorld/resources/cross-origin-xhr.txt' because it violates the following Content Security Policy directive: "connect-src 'none'".

CONSOLE ERROR: line 113: Refused to connect to 'http://localhost:8000/security/isolatedWorld/resources/cross-origin-xhr.txt' because it violates the following Content Security Policy directive: "connect-src 'none'".

Tests that isolated worlds can have XHRs that the page's CSP wouldn't allow.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

XHR from main world
PASS: The request was disallowed
XHR from isolated world with unchanged CSP
PASS: The request was disallowed
XHR from isolated world with same security origin as XHR target.
PASS: The request was disallowed
XHR from isolated world with same security origin as XHR target, and looser CSP.
PASS: The request succeeded
XHR from main world is not affected by the isolated world origin or CSP
PASS: The request was disallowed
PASS successfullyParsed is true

TEST COMPLETE
